Fixed incorrect vintique sound website link, simplified theme context code into a single file.
