+++ src/context/theme-context.ts
@@ -1,15 +1,23 @@
-import { createContext } from "react"
+import { createContext, useContext } from "react"
 
-export type Mode = "light" | "dark" | "system"
+export type ThemeMode = "light" | "dark" | "system"
 
 export interface ThemeContextType {
   permIndex: number
   setPermIndex: (index: number) => void
   isDark: boolean
-  theme: Mode
-  setTheme: (theme: Mode) => void
+  theme: ThemeMode
+  setTheme: (theme: ThemeMode) => void
   next: () => void
   prev: () => void
 }
 
 export const ThemeContext = createContext<ThemeContextType | undefined>(undefined)
+
+export function useTheme(): ThemeContextType {
+  const context = useContext(ThemeContext)
+  if (context === undefined) {
+    throw new Error("useTheme must be used within a ThemeProvider")
+  }
+  return context
+}
